+++ src/filters/FactionLegend.css
@@ -75,47 +75,47 @@ div.faction-legend .faction .logo.all-factions {
 }
 
 div.faction-legend .faction .logo.ktb {
-  background-image: url('./imgs/logos/ktb.svg');
+  background-image: url('../imgs/logos/ktb.svg');
 }
 
 div.faction-legend .faction .logo.union {
-  background-image: url('./imgs/logos/union.svg');
+  background-image: url('../imgs/logos/union.svg');
 }
 
 div.faction-legend .faction .logo.ha {
-  background-image: url('./imgs/logos/ha.svg');
+  background-image: url('../imgs/logos/ha.svg');
 }
 
 div.faction-legend .faction .logo.aun {
-  background-image: url('./imgs/logos/aun.svg');
+  background-image: url('../imgs/logos/aun.svg');
 }
 
 div.faction-legend .faction .logo.albatross {
-  background-image: url('./imgs/logos/albatross.svg');
+  background-image: url('../imgs/logos/albatross.svg');
 }
 
 div.faction-legend .faction .logo.gms {
-  background-image: url('./imgs/logos/gms.svg');
+  background-image: url('../imgs/logos/gms.svg');
 }
 
 div.faction-legend .faction .logo.horus {
-  background-image: url('./imgs/logos/horus.svg');
+  background-image: url('../imgs/logos/horus.svg');
 }
 
 div.faction-legend .faction .logo.ipsn {
-  background-image: url('./imgs/logos/ipsn.svg');
+  background-image: url('../imgs/logos/ipsn.svg');
 }
 
 div.faction-legend .faction .logo.sparri {
-  background-image: url('./imgs/logos/sparri.svg');
+  background-image: url('../imgs/logos/sparri.svg');
 }
 
 div.faction-legend .faction .logo.ssc {
-  background-image: url('./imgs/logos/ssc.svg');
+  background-image: url('../imgs/logos/ssc.svg');
 }
 
 div.faction-legend .faction .logo.voladores {
-  background-image: url('./imgs/logos/voladores.svg');
+  background-image: url('../imgs/logos/voladores.svg');
 }
 
 div.source-legend .source .sname {
