+++ src/filters/FactionLegend.js
