+++ src/filters/PublishingAttributeLegend.js
