+++ src/source-table/SourceTableEntry.js
