+++ src/timeline/CopyLink.js
@@ -1,5 +1,5 @@
 import { useState, useContext } from 'react';
-import { ActionMessageContext } from './ActionMessageContext';
+import { ActionMessageContext } from '../ActionMessageContext';
 
 function CopyLink({ elementId, name }) {
     const showActionMessage = useContext(ActionMessageContext);
