+++ src/timeline/TimelineEntry.css
@@ -51,15 +51,15 @@
 }
 
 .timeline ul li div.shareable-link {
-    background-image: url('./imgs/link-icon-symbol.svg');
+    background-image: url('../imgs/link-icon-symbol.svg');
 }
 
 .timeline ul li div.modify-entry {
-    background-image: url('./imgs/modify-icon.svg');
+    background-image: url('../imgs/modify-icon.svg');
 }
 
 .timeline ul li div.remove-entry {
-    background-image: url('./imgs/remove-icon.svg');
+    background-image: url('../imgs/remove-icon.svg');
 }
 
 .timeline ul li div.shareable-link:hover,
@@ -212,47 +212,47 @@ li.timeline-entry .factions span {
 }
 
 li.timeline-entry .factions .ktb {
-    background-image: url('./imgs/logos/ktb.svg');
+    background-image: url('../imgs/logos/ktb.svg');
 }
 
 li.timeline-entry .factions .union {
-    background-image: url('./imgs/logos/union.svg');
+    background-image: url('../imgs/logos/union.svg');
 }
 
 li.timeline-entry .factions .ha {
-    background-image: url('./imgs/logos/ha.svg');
+    background-image: url('../imgs/logos/ha.svg');
 }
 
 li.timeline-entry .factions .aun {
-    background-image: url('./imgs/logos/aun.svg');
+    background-image: url('../imgs/logos/aun.svg');
 }
 
 li.timeline-entry .factions .albatross {
-    background-image: url('./imgs/logos/albatross.svg');
+    background-image: url('../imgs/logos/albatross.svg');
 }
 
 li.timeline-entry .factions .gms {
-    background-image: url('./imgs/logos/gms.svg');
+    background-image: url('../imgs/logos/gms.svg');
 }
 
 li.timeline-entry .factions .horus {
-    background-image: url('./imgs/logos/horus.svg');
+    background-image: url('../imgs/logos/horus.svg');
 }
 
 li.timeline-entry .factions .ipsn {
-    background-image: url('./imgs/logos/ipsn.svg');
+    background-image: url('../imgs/logos/ipsn.svg');
 }
 
 li.timeline-entry .factions .sparri {
-    background-image: url('./imgs/logos/sparri.svg');
+    background-image: url('../imgs/logos/sparri.svg');
 }
 
 li.timeline-entry .factions .ssc {
-    background-image: url('./imgs/logos/ssc.svg');
+    background-image: url('../imgs/logos/ssc.svg');
 }
 
 li.timeline-entry .factions .voladores {
-    background-image: url('./imgs/logos/voladores.svg');
+    background-image: url('../imgs/logos/voladores.svg');
 }
 
 li.edited .date span::after {
