+++ src/timeline/TimelineEntrySource.js
